[core] move OverlayInstance to separate module; make all handlers optional
